feat: modify Info.VmCount uint32 -> int64

+++ node/schema/schema.go
@@ -14,7 +14,7 @@ type Info struct {
 	Token       string              `json:"Token"`
 	Registry    string              `json:"Registry"`
 	Node        registrySchema.Node `json:"Node"`
-	VmCount     uint32              `json:"Vm-Count"`
+	VmCount     int64               `json:"Vm-Count"`
 }
 
 type ResultHandler func(vmmSchema.Result)
